Refactor all transformation and directory code to Batch class and subclasses
Completely refactor the active code for transforming input data sets into sets for ingest.
